Change type of permissions value from int to long for 128 bit rc4 passwords (http://stackoverflow.com/questions/24391307/itextsharp-cannot-open-files-with-password-if-128-bit-rc4)
git-svn-id: https://svn.code.sf.net/p/itext/code/trunk@6491 3a709846-db2a-4fbd-be0b-eb932eee6479

+++ itext/src/main/java/com/itextpdf/text/pdf/PdfEncryption.java
@@ -116,7 +116,7 @@ public class PdfEncryption {
 	/** The public key security handler for certificate encryption */
 	protected PdfPublicKeySecurityHandler publicKeyHandler = null;
 
-	int permissions;
+	long permissions;
 
 	byte documentID[];
 
@@ -209,7 +209,7 @@ public class PdfEncryption {
 		return encryptMetadata;
 	}
 
-    public int getPermissions() {
+    public long getPermissions() {
         return permissions;
     }
 
@@ -270,7 +270,7 @@ public class PdfEncryption {
 	 * ownerKey, documentID must be setup
 	 */
 	private void setupGlobalEncryptionKey(byte[] documentID, byte userPad[],
-			byte ownerKey[], int permissions) {
+			byte ownerKey[], long permissions) {
 		this.documentID = documentID;
 		this.ownerKey = ownerKey;
 		this.permissions = permissions;
@@ -493,7 +493,7 @@ public class PdfEncryption {
 	/**
 	 */
 	public void setupByUserPassword(byte[] documentID, byte userPassword[],
-			byte ownerKey[], int permissions) {
+			byte ownerKey[], long permissions) {
 		setupByUserPad(documentID, padPassword(userPassword), ownerKey,
 				permissions);
 	}
@@ -501,7 +501,7 @@ public class PdfEncryption {
 	/**
 	 */
 	private void setupByUserPad(byte[] documentID, byte userPad[],
-			byte ownerKey[], int permissions) {
+			byte ownerKey[], long permissions) {
 		setupGlobalEncryptionKey(documentID, userPad, ownerKey, permissions);
 		setupUserKey();
 	}
@@ -509,13 +509,13 @@ public class PdfEncryption {
 	/**
 	 */
 	public void setupByOwnerPassword(byte[] documentID, byte ownerPassword[],
-			byte userKey[], byte ownerKey[], int permissions) {
+			byte userKey[], byte ownerKey[], long permissions) {
 		setupByOwnerPad(documentID, padPassword(ownerPassword), userKey,
 				ownerKey, permissions);
 	}
 
 	private void setupByOwnerPad(byte[] documentID, byte ownerPad[],
-			byte userKey[], byte ownerKey[], int permissions) {
+			byte userKey[], byte ownerKey[], long permissions) {
 		byte userPad[] = computeOwnerKey(ownerKey, ownerPad); // userPad will
 																// be set in
 																// this.ownerKey
